pyproject.toml
[project]
name = "sillyAgent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.9",
    "huggingface-hub>=1.5.0",
    "langchain-neo4j>=0.8.0",
    "langchain-openai>=1.1.10",
    "mem0ai>=1.0.4",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.3",
    "config>=0.5.1",
    "tzdata>=2025.3",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "pydantic>=2.12.5",
    "websockets>=16.0",
    "colorlog>=6.10.1",
]
